Harness: add a crash, improve docs

+++ harness/library.h
@@ -1,5 +1,5 @@
 // an 'nprintf' implementation - print the first len bytes of data
-void lib_echo(char *data, int len);
+void lib_echo(char *data, ssize_t len);
 
 // optimised multiply - returns x*y
 int  lib_mul(int x, int y);
